requirements.txt
burnysc2==5.0.4
numpy==1.19.3 # 1.19.4 is broken on Windows https://developercommunity.visualstudio.com/content/problem/1207405/fmod-after-an-update-to-windows-2004-is-causing-a.html

# Dev dependencies
black==20.8b1
flake8==3.8.4
mypy==0.790
pre-commit==2.8.2
